// Fabrikage/Fabrikage.CoreProtocols/src/types.ts
export interface SchemaDefinition extends JSONSchema7 {
  $id: string;
  $schema?: string;
  title: string;
  description?: string;
  type: string | string[];
  definitions?: Record<string, JSONSchema7>;
  [key: string]: any;
}